pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "afrorad-whisper-pipeline"
version = "0.1.0"
description = "Hydra + Accelerate pipeline for Whisper fine-tuning/evaluation on AfroRadVoice-FR"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "AfroRad Team" }
]
dependencies = [
  "accelerate>=1.2.0",
  "datasets[audio]>=3.1.0",
  "evaluate>=0.4.3",
  "hydra-core>=1.3.2",
  "jiwer>=3.0.4",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "omegaconf>=2.3.0",
  "pandas>=2.2.0",
  "soundfile>=0.12.1",
  "torch>=2.2.0",
  "torchaudio>=2.2.0",
  "transformers>=4.46.0",
  "huggingface-hub>=0.28.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.9.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
